refactor: 수정한 EmptyState 적용(Next-Study-habjung#72)
