+++ src/components/profile/container/ProductListContainer.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 
+import EmptyState from '@/components/common/empty-state/EmptyState';
 import Text from '@/components/common/text/Text';
 import useProductList from '@/queries/product/useProductList';
 
-import EmptyState from '../components/empty-state/EmptyState';
 import ProductCard from '../components/product/ProductCard';
 import { useProduct } from '../hook/useProduct';
 
@@ -25,7 +25,10 @@ export default function ProductListContainer({
         판매중인 상품
       </Text>
       {productList?.data === 0 ? (
-        <EmptyState className="mx-4" message="현재 판매중인 상품이 없습니다." />
+        <EmptyState
+          className="mx-4"
+          description="현재 판매중인 상품이 없습니다."
+        />
       ) : (
         <div className="horizontal-scroll scrollbar-hidden mt-4 flex gap-15px px-4">
           {productList?.product.map((productData) => (
